backend/requirements.txt
# AI Focus Groups - Backend Production Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
authlib==1.3.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# AI Integration (will be added in Phase 4)
# openai==1.10.0
# anthropic==0.8.1

# Image Generation (will be added in Phase 4)
# pillow==10.2.0

# Caching (will be added in Phase 9)
# redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
